Added dotnet sonar scanner. Refactored to share base sonar scanner dependencies.
